requirements.txt
# Backend
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sse-starlette>=3.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
python-multipart>=0.0.18
aiosqlite>=0.20.0
pyyaml>=6.0

# LLM Providers
openai>=1.60.0
anthropic>=0.42.0
google-genai>=1.0.0

# RAG
chromadb>=0.5.0
pypdf>=5.0.0

# Frontend
streamlit>=1.41.0
httpx>=0.28.0
httpx-sse>=0.4.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
